requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6


SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2   # for migrations (optional but recommended)

# --- Data Validation ---
pydantic==2.8.2
pydantic-settings==2.3.3

# --- Auth & Security ---
python-jose[cryptography]==3.3.0

# --- Environment Variables ---
python-dotenv==1.0.1

# --- Testing ---
pytest==8.3.2
httpx==0.27.2

loguru==0.7.2 

psycopg2-binary==2.9.9
authlib==1.3.2
PyJWT==2.9.0

python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.2   # when you add Google SSO later
email-validator==2.3.0
python-multipart==0.0.20
bcrypt-5.0.0
google-auth=2.23.0
google-auth-oauthlib=1.0.0
google-auth-httplib2=0.1.0
